feat(converter): refactor FiatConverter to debounce and typed
